skills/external-review/server/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
pyyaml>=6.0

# Dev/Test
pytest>=8.0.0
pytest-asyncio>=0.24.0
